requirements.txt
pandas
matplotlib
seaborn
fpdf
Pillow
